add more fixes to msvc build (#7468)
* add more fixes to msvc build

* more fixes

+++ libr/asm/arch/nios/gnu/nios2-dis.c
@@ -254,7 +254,7 @@ nios2_print_insn_arg (const char *argptr,
       break;
     case 'i':
       /* 16-bit signed immediate.  */
-      i = (signed) (GET_INSN_FIELD (IMM16, opcode) << 16) >> 16;
+      i = (int) (GET_INSN_FIELD (IMM16, opcode) << 16) >> 16;
       (*info->fprintf_func) (info->stream, "%ld", i);
       break;
     case 'u':
@@ -264,7 +264,7 @@ nios2_print_insn_arg (const char *argptr,
       break;
     case 'o':
       /* 16-bit signed immediate address offset.  */
-      i = (signed) (GET_INSN_FIELD (IMM16, opcode) << 16) >> 16;
+      i = (int) (GET_INSN_FIELD (IMM16, opcode) << 16) >> 16;
       address = address + 4 + i;
       (*info->print_address_func) (address, info);
       break;
